requirements.txt
altgraph==0.17.3
aniso8601==9.0.1
auto-py-to-exe==2.29.0
bcrypt==3.2.0
bottle==0.12.23
bottle-websocket==0.2.9
capitalize==1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.1.3
colorama==0.4.6
cryptography==36.0.1
curve==0.0.7
cycler==0.11.0
DateTime==5.1
DAWG-Python==0.7.2
decorator==5.1.1
docopt==0.6.2
docxcompose==1.3.4
docxtpl==0.15.2
Eel==0.14.0
et-xmlfile==1.1.0
fabric==2.6.0
Flask==2.2.3
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-WTF==1.1.1
fonttools==4.29.1
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.3
invoke==1.6.0
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.0.3
kiwisolver==1.3.2
line==0.8.2
lxml==4.8.0
MarkupSafe==2.1.2
matplotlib==3.5.1
numpy==1.22.2
openpyxl==3.1.1
packaging==21.3
paramiko==2.8.1
pathlib2==2.3.6
pefile==2023.2.7
Pillow==9.0.1
possible==0.0.0
pyasn1==0.4.8
pybind11==2.9.0
pycparser==2.21
pygame==2.1.2
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyNaCl==1.4.0
pyparsing==3.0.7
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2022.7.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
row==0.1.0.dev3
rsa==4.8
scipy==1.8.0
self==2020.12.3
six==1.16.0
SQLAlchemy==2.0.4
SQLAlchemy-serializer==1.4.1
turtle==0.0.1
typing_extensions==4.5.0
urllib3==1.26.8
Werkzeug==2.2.3
whichcraft==0.6.1
WTForms==3.0.1
XlsxWriter==3.0.3
zope.event==4.6
zope.interface==5.5.2
